add user api auth access

+++ project_api_car/project_api_car/src/main/java/com/example/project_api_car/dto/IBaseDto.java
@@ -12,5 +12,5 @@ public abstract class IBaseDto {
     private String Database;
     private String UpdatedBy;
     private Date UpdatedDate;
-    private Long RecordCount;
+    private int RecordCount;
 }
